Cleanup of logger API
