+++ opencog/guile/LoggerSCM.cc
@@ -22,12 +22,38 @@
  */
 
 #include <opencog/util/Logger.h>
+#include <opencog/guile/SchemeModule.h>
 #include "SchemePrimitive.h"
-#include "LoggerSCM.h"
 
 using namespace opencog;
-
 namespace opencog {
+
+/**
+ * Expose the Logger singleton to Scheme
+ */
+
+class LoggerSCM : public ModuleWrap
+{
+protected:
+	virtual void init();
+
+	std::string do_logger_set_level(const std::string& level);
+	std::string do_logger_get_level(void);
+	std::string do_logger_set_filename(const std::string& filename);
+	std::string do_logger_get_filename(void);
+	bool do_logger_set_stdout(bool);
+	bool do_logger_set_sync(bool);
+	bool do_logger_set_timestamp(bool);
+	void do_logger_error(const std::string& msg);
+	void do_logger_warn(const std::string& msg);
+	void do_logger_info(const std::string& msg);
+	void do_logger_debug(const std::string& msg);
+	void do_logger_fine(const std::string& msg);
+
+public:
+	LoggerSCM();
+};
+
 
 /// Set level, return previous level.
 std::string LoggerSCM::do_logger_set_level(const std::string& level)
@@ -139,6 +165,10 @@ void LoggerSCM::init(void)
 	define_scheme_primitive("cog-logger-fine-str",
 		&LoggerSCM::do_logger_fine, this, "logger");
 }
+
+extern "C" {
+void opencog_logger_init(void);
+};
 
 void opencog_logger_init(void)
 {
